revamped and more comprehensive logic for receipt creation and dependents. BookDetail split into supply price and sales price. TODO: generate new data and test

+++ src/main/java/sd_009/bookstore/dto/jsonApiResource/receipt/PaymentDetailDto.java
@@ -26,5 +26,5 @@ public class PaymentDetailDto implements Serializable {
     private final PaymentType paymentType;
     private final String provider;
     private final String providerId;
-    private final Long amount;
+    private final Double amount;
 }
